Refactor _pareto_frontier to return indices.
The `_pareto_frontier` function now returns the indices of the points on the Pareto frontier. This enables `_get_tn_fn_counts` to also return the thresholds associated with the frontier points. The JAX implementation has been adjusted to operate on indices.

This lets us maintain the threshold values on the frontier, which will be used for statistically accurate epsilon lower bounds computations. We will choose an optimal threshold on a subset of the data, and then estimate epsilon on the complement.

PiperOrigin-RevId: 839505405